refactor: renamed firebaseUtils to dbUtils and added comments
