+++ frontend/pages/index.vue
@@ -5,7 +5,7 @@
 <script lang="ts">
 import 'reflect-metadata'
 import { Vue, Component } from 'vue-property-decorator'
-import { getRecentReleases, getCategories, getProjectConfigs } from '../assets/firebaseUtils'
+import { getRecentReleases, getCategories, getProjectConfigs } from '../assets/dbUtils'
 import Homepage from '@/components/Homepage/index.vue'
 
 // @ts-ignore - idk why this shows "no overload matches this call..."
