+++ frontend/pages/platform/_platformId/index.vue
@@ -6,7 +6,7 @@
 import 'reflect-metadata'
 import { Vue, Component } from 'vue-property-decorator'
 
-import { getCategories, getProjectConfigs } from '../../../assets/firebaseUtils'
+import { getCategories, getProjectConfigs } from '../../../assets/dbUtils'
 import Homepage from '@/components/Homepage/index.vue'
 
 // @ts-ignore - idk why this shows "no overload matches this call..."
